backend/icycon/requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
async-timeout==5.0.1
backcall==0.2.0
billiard==4.2.2
boto3==1.40.59
botocore==1.40.59
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
connectors==4.1
decorator==5.2.1
Django==4.2.23
dj-database-url==0.5.0 
django-cors-headers==4.9.0
django-extensions==4.1
django-tailwind==3.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
docx==0.2.4
exceptiongroup==1.3.0
executing==2.2.0
graphviz==0.21
idna==3.11
ipython==8.15.0
jedi==0.19.2
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pickleshare==0.7.5
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
xlsxwriter==3.2.9
whitenoise
gunicorn
openai==1.51.0
